Long-due update
A.o.
- Rewrite ComplexInstr to add its instructions UPON RUNNING, and only
if they're not already part of #bod
- Save the timestamps of the first display of the page (handle preload)
< preload = only thing not done yet (though first run)
and the validation of the trial (finishedCallback)
- Make it possible to check for a subset of preloads (item-label-based)
> also revise the preloading strategy:
right now, callback instructions (for instance) don't add their
resources to 'toPreload'
- Make compatible with GetItemsFrom / Rewrite GetItemsFrom as
PennController.FeedItems
- Add a Finish/Done/End/… Instruction to (prematurely) go to the next
trial/to the end?
- Add something to check the ACTUAL start-playing of audio/video
- Add a t.selector instruction
can click on each instruction's element or use a key (how should the
keys be set?)
can be just once or as many times
function to enable/disable
callback function with instruction?
- Add a way to identify the selected answer in SELECTOR, when shuffled
- Rewrite AUDIOs so they behave the same as RADIOs
they record every event for themselves and user decides which to store
with the SAVE method